+++ src/compas_fab/backends/ros/backend_features/move_it_plan_cartesian_motion.py
@@ -2,6 +2,8 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import itertools
+
 from compas.utilities import await_callback
 
 from compas_fab.backends.interfaces import PlanCartesianMotion
@@ -112,12 +114,12 @@ class MoveItPlanCartesianMotion(PlanCartesianMotion):
         joint_state = JointState(header=header,
                                  name=start_configuration.joint_names,
                                  position=start_configuration.values)
-        start_state = RobotState(joint_state, MultiDOFJointState(header=header))
+        start_state = RobotState(joint_state, MultiDOFJointState(header=header), is_diff=True)
 
-        if options.get('attached_collision_meshes'):
-            for acm in options['attached_collision_meshes']:
-                aco = AttachedCollisionObject.from_attached_collision_mesh(acm)
-                start_state.attached_collision_objects.append(aco)
+        attached_collision_meshes = options.get('attached_collision_meshes', [])
+        for acm in attached_collision_meshes:
+            aco = AttachedCollisionObject.from_attached_collision_mesh(acm)
+            start_state.attached_collision_objects.append(aco)
 
         path_constraints = convert_constraints_to_rosmsg(options.get('path_constraints'), header)
 
@@ -145,6 +147,9 @@ class MoveItPlanCartesianMotion(PlanCartesianMotion):
                 start_state = response.start_state.joint_state
                 start_state_types = [joint_type_by_name[name] for name in start_state.name]
                 trajectory.start_configuration = Configuration(start_state.position, start_state_types, start_state.name)
+                trajectory.attached_collision_meshes = list(itertools.chain(*[
+                    aco.to_attached_collision_meshes()
+                    for aco in response.start_state.attached_collision_objects]))
 
                 callback(trajectory)